ai_service/requirements.txt
flask
flask-cors
python-dotenv
transformers
torch